+++ RayTracer/src/cameras.h
@@ -27,7 +27,7 @@ public:
 	/* Ray Tracing params */
 	int max_depth = 10; /* Maximum number of bounces per ray */
 	unsigned int current_samples = 0; /* Used to determine previous sample contributions to the accumulated image */
-	std::vector<float> image_accumulator; /* Stores results from all previous samples used for accumulation */
+	std::vector<double> image_accumulator; /* Stores results from all previous samples used for accumulation */
 	bool simulate_time = false; /* Determines if camera has a "shutter speed" to simulate effects like motion blur.
 								   Note: Timescale for the cameras is always defined within 0-1; it is up to the user
 								   to decide how much/where objects move within that time frame. */
@@ -36,7 +36,7 @@ public:
 	bool gamma_correct = false; /* OpenGL gamma corrects for us so this is optional */
 
 protected:
-	float aspect_ratio = 1.0f; /* Ratio of image width over image height */
+	double aspect_ratio = 1.0; /* Ratio of image width over image height */
 
 	/* Store previous sample's view params */
 	unsigned int old_image_width = 100;
@@ -48,9 +48,9 @@ class ProjectiveCamera : public Camera
 {
 public:
 	/* View Params */
-	Point3 origin = Point3(0.0f, 0.0f, 0.0f); /* Point the camera is looking from */
-	Point3 look_at = Point3(0.0f, 0.0f, -1.0f); /* Point the camera is looking at */
-	Vec3 up = Vec3(0.0f, 1.0f, 0.0f); /* Camera-relative "up" vector */
+	Point3 origin = Point3(0.0, 0.0, 0.0); /* Point the camera is looking from */
+	Point3 look_at = Point3(0.0, 0.0, -1.0); /* Point the camera is looking at */
+	Vec3 up = Vec3(0.0, 1.0, 0.0); /* Camera-relative "up" vector */
 
 protected:
 	/* Calculated params (in Initialize) */
@@ -77,10 +77,10 @@ public:
 	void Initialize() override;
 
 public:
-	float vfov = 90.0f; /* Vertical field of view */
+	double vfov = 90.0; /* Vertical field of view */
 
 private:
-	float old_vfov; /* Store the previous sample's vfov */
+	double old_vfov; /* Store the previous sample's vfov */
 
 };
 
@@ -92,9 +92,9 @@ public:
 	void Initialize();
 
 public:
-	float defocus_angle = 0.0f; /* Variation angle of rays through each pixel */
-	float focus_distance = 10.0f; /* Distance from the camera origin to plane of perfect focus */
-	float vfov = 90.0f; /* Vertical field of view */
+	double defocus_angle = 0.0; /* Variation angle of rays through each pixel */
+	double focus_distance = 10.0; /* Distance from the camera origin to plane of perfect focus */
+	double vfov = 90.0; /* Vertical field of view */
 
 private:
 	/* Returns a random point in the camera defocus disk */
@@ -105,7 +105,7 @@ private:
 	Vec3 defocus_disk_u; /* Defocus disk horizontal radius */
 	Vec3 defocus_disk_v; /* Defocus disk vertical radius */
 
-	float old_vfov; /* Store the previous sample's vfov */
+	double old_vfov; /* Store the previous sample's vfov */
 };
