+++ RayTracer/src/material.h
@@ -38,30 +38,30 @@ private:
 class Metal : public Material
 {
 public:
-	Metal(const Color& albedo, float roughness) : albedo(albedo), roughness(roughness < 1.0f ? roughness : 1.0f) {}
+	Metal(const Color& albedo, double roughness) : albedo(albedo), roughness(roughness < 1.0 ? roughness : 1.0) {}
 
 	bool Scatter(const Ray& ray_in, const Interaction& interaction, Color& attenuation, Ray& ray_out) const override;
 
 private:
 	Color albedo;
-	float roughness; /* "Fuzziness" of the reflection -- 0 = perfect reflection, 1 = diffuse */
+	double roughness; /* "Fuzziness" of the reflection -- 0 = perfect reflection, 1 = diffuse */
 };
 
 
 class Dielectric : public Material
 {
 public:
-	Dielectric(float eta_out, float eta_in) : eta_out(eta_out), eta_in(eta_in) {}
-	Dielectric(float eta_in_over_out) : eta_out(1.0f), eta_in(eta_in_over_out) {}
+	Dielectric(double eta_out, double eta_in) : eta_out(eta_out), eta_in(eta_in) {}
+	Dielectric(double eta_in_over_out) : eta_out(1.0), eta_in(eta_in_over_out) {}
 
 	bool Scatter(const Ray& ray_in, const Interaction& interaction, Color& attenuation, Ray& ray_out) const override;
 
 private:
-	float eta_out; /* Refractive index of the enclosing media */
-	float eta_in; /* Refractive index of the material */
+	double eta_out; /* Refractive index of the enclosing media */
+	double eta_in; /* Refractive index of the material */
 
 	/* Use Schlick's approximation to model reflectance */
-	static float Reflectance(float cosine, float refraction_index);
+	static double Reflectance(double cosine, double refraction_index);
 };
 
 } /* namespace rt */
