+++ RayTracer/src/math.h
@@ -14,13 +14,13 @@ namespace rt
 /* === Accessor redefinitions === */
 /* ============================== */
 
-using Vec2 = glm::vec2;
-using Point2 = glm::vec2;
-using Vec3 = glm::vec3;
-using Point3 = glm::vec3;
-using Color = glm::vec3;
-using Vec4 = glm::vec4;
-using Mat4 = glm::mat4;
+using Vec2 = glm::dvec2;
+using Point2 = glm::dvec2;
+using Vec3 = glm::dvec3;
+using Point3 = glm::dvec3;
+using Color = glm::dvec3;
+using Vec4 = glm::dvec4;
+using Mat4 = glm::dmat4;
 
 
 /* ============= */
@@ -38,37 +38,37 @@ inline bool Equal(T u, T v)
 template <typename T>
 inline bool NearZero(T v)
 {
-	return glm::all(glm::epsilonEqual(v, T(0.0f), Eps));
+	return glm::all(glm::epsilonEqual(v, T(0.0), Eps));
 }
 
-/* Generates a Vec3 with all components between 0.0f and 1.0f */
+/* Generates a Vec3 with all components between 0.0 and 1.0 */
 inline Vec3 RandomVec3()
 {
-	return Vec3(RandomFloat(), RandomFloat(), RandomFloat());
+	return Vec3(RandomDouble(), RandomDouble(), RandomDouble());
 }
 
 /* Generates a Vec3 with all components between provided min and max values */
 inline Vec3 RandomVec3(float min, float max)
 {
-	return Vec3(RandomFloat(min, max), RandomFloat(min, max), RandomFloat(min, max));
+	return Vec3(RandomDouble(min, max), RandomDouble(min, max), RandomDouble(min, max));
 }
 
-/* Generates a Vec2 with length < 1.0f */
+/* Generates a Vec2 with length < 1.0 */
 inline Vec2 RandomInUnitDisk()
 {
 	while (true)
 	{
-		Vec2 p = Vec2(RandomFloat(-1.0f, 1.0f), RandomFloat(-1.0f, 1.0f));
-		if (glm::length2(p) < 1) return p;
+		Vec2 p = Vec2(RandomDouble(-1.0, 1.0), RandomDouble(-1.0, 1.0));
+		if (glm::length2(p) < 1.0) return p;
 	}
 }
 
-/* Generate a random Vec3 with a length < 1.0f */
+/* Generate a random Vec3 with a length < 1.0 */
 inline Vec3 RandomInUnitSphere()
 {
 	while (true)
 	{
-		auto p = RandomVec3(-1.0f, 1.0f);
+		auto p = RandomVec3(-1.0, 1.0);
 		if (glm::length2(p) < 1) return p;
 	}
 }
@@ -85,21 +85,21 @@ inline Vec3 RandomOnHemisphere(const Vec3& normal)
 	Vec3 on_unit_sphere = RandomUnitVector();
 
 	/* In the same hemisphere as the normal? */
-	if (glm::dot(on_unit_sphere, normal) > 0.0f) return on_unit_sphere;
+	if (glm::dot(on_unit_sphere, normal) > 0.0) return on_unit_sphere;
 	else return -on_unit_sphere;
 }
 
 
-/* Returns a random point in the [-0.5f, -0.5f] to [0.5f, 0.5f] unit square */
+/* Returns a random point in the [-0.5, -0.5] to [0.5, 0.5] unit square */
 inline Point2 SampleSquare()
 {
-	return Point2(RandomFloat() - 0.5f, RandomFloat() - 0.5f);
+	return Point2(RandomDouble() - 0.5, RandomDouble() - 0.5);
 }
 
 /* Reflect provided Vec3 along the provided normal */
 inline Vec3 Reflect(const Vec3& v, const Vec3& n)
 {
-	return v - 2 * glm::dot(v, n) * n;
+	return v - 2.0 * glm::dot(v, n) * n;
 }
 
 /* Refract the (Vec3) incident ray direction along the provided normal using the provided index of refraction ratio */
